feat: improve xbf2 external font metrics and rendering
Add xbf2-aware scanning, 16-bit advance handling, binary-search metrics lookup, and baseline-aware rendering so external fonts render with correct spacing and layout.

+++ lib/ExternalFont/ExternalFont.h
@@ -8,7 +8,7 @@
 struct ExternalGlyphMetrics {
   uint8_t width = 0;
   uint8_t height = 0;
-  uint8_t advanceX = 0;
+  uint16_t advanceX = 0;
   uint16_t flags = 0;
   int16_t left = 0;
   int16_t top = 0;
